Refactors the code by introducing shared attribute constructors
The goal of this refactor was to improve the generated documentation by providing clearer attribute descriptions and re-organizing common attributes, which was not the case earlier. This approach also follows the DRY principle making the code cleaner and easier to maintain, as now we centralize the attribute metadata definition. The changes also involve the update of test files due to compatibility reasons after the refactor.